requirements.txt
appJar
numpy
opencv-python
Pillow
